third_party/blink/web_tests/plugins/mouse-move-over-plugin-in-frame-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
Test for bug 47956: Crash when mouse is over plugin inside a frame.
